Extract helper methods into FlowUnitTestHelper module
I plan to re-use these methods in another test.
